Remove the instruction-by-instruction inspector
Get rid of one the epicenters of needless uncontrollable
IO done by the machine. This can and should be done by
the CLI / some other kind of debug tool that implements a
special callback which, if set, causes the machine to halt
after every step. A protocolized callback return value may
be used to communicate the contination intent to the machine.

This ends an era!